api/requirements.txt
alembic==1.12.0 ; python_version >= "3.10" and python_version < "4.0"
annotated-types==0.5.0 ; python_version >= "3.10" and python_version < "4.0"
anyio==3.7.1 ; python_version >= "3.10" and python_version < "4.0"
black==23.9.1 ; python_version >= "3.10" and python_version < "4.0"
boto3==1.28.59 ; python_version >= "3.10" and python_version < "4.0"
botocore==1.31.59 ; python_version >= "3.10" and python_version < "4.0"
certifi==2023.7.22 ; python_version >= "3.10" and python_version < "4.0"
charset-normalizer==3.3.0 ; python_version >= "3.10" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.10" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "4.0" and platform_system == "Windows"
exceptiongroup==1.1.3 ; python_version >= "3.10" and python_version < "3.11"
fastapi==0.103.2 ; python_version >= "3.10" and python_version < "4.0"
greenlet==3.0.0 ; python_version >= "3.10" and python_version < "4.0" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32")
h11==0.14.0 ; python_version >= "3.10" and python_version < "4.0"
idna==3.4 ; python_version >= "3.10" and python_version < "4.0"
jmespath==1.0.1 ; python_version >= "3.10" and python_version < "4.0"
mako==1.2.4 ; python_version >= "3.10" and python_version < "4.0"
markupsafe==2.1.3 ; python_version >= "3.10" and python_version < "4.0"
mypy-extensions==1.0.0 ; python_version >= "3.10" and python_version < "4.0"
packaging==23.2 ; python_version >= "3.10" and python_version < "4.0"
pathspec==0.11.2 ; python_version >= "3.10" and python_version < "4.0"
pillow==10.0.1 ; python_version >= "3.10" and python_version < "4.0"
platformdirs==3.11.0 ; python_version >= "3.10" and python_version < "4.0"
psycopg2==2.9.9 ; python_version >= "3.10" and python_version < "4.0"
pydantic-core==2.10.1 ; python_version >= "3.10" and python_version < "4.0"
pydantic-settings==2.0.3 ; python_version >= "3.10" and python_version < "4.0"
pydantic==2.4.2 ; python_version >= "3.10" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.10" and python_version < "4.0"
python-dotenv==1.0.0 ; python_version >= "3.10" and python_version < "4.0"
python-multipart==0.0.6 ; python_version >= "3.10" and python_version < "4.0"
requests==2.31.0 ; python_version >= "3.10" and python_version < "4.0"
s3transfer==0.7.0 ; python_version >= "3.10" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.10" and python_version < "4.0"
sniffio==1.3.0 ; python_version >= "3.10" and python_version < "4.0"
sqlalchemy==2.0.21 ; python_version >= "3.10" and python_version < "4.0"
starlette==0.27.0 ; python_version >= "3.10" and python_version < "4.0"
tomli==2.0.1 ; python_version >= "3.10" and python_version < "3.11"
typing-extensions==4.8.0 ; python_version >= "3.10" and python_version < "4.0"
urllib3==1.26.17 ; python_version >= "3.10" and python_version < "4.0"
uvicorn==0.23.2 ; python_version >= "3.10" and python_version < "4.0"
